Improve abort behavior and documentation.
"Abort" has been replaced by unconditional immediate termination.
`libc::abort()` terminates via the SIGABRT signal, which can be caught
by processes. Using SIGKILL is uncatchable, and an equivalent "FailFast"
exception is used on Windows.

The reasoning behind this is to control the blast area for floating
point exceptions. SIGFPE is considered unrecoverable, and `batman` does
its best to ensure that is consistent even in the face of code that
would attempt to thwart the guarantee of unrecoverability. This is
codified in the SAFETY invariants for `batman::signal()`.

Debuggers get a free pass here because they can generally do _anything_
to a process, including taking over signal handlers or even changing
arbitrary state (or code!) at runtime. None of that contributes to UB.
So, debuggers do not factor into safety invariants.

On the documentation side, a lot of attention is paid to how things
could go wrong in the signal handler. There are some open questions
remaining (marked with "TODO" comments) but otherwise I believe
everything else that could be known about the signal handler is
described clearly.